update src/pages for split PromptDashboard to router pages
