+++ src/pages/PromptDashboard.jsx
@@ -1,12 +1,11 @@
 // PromptDashboard.jsx
 import * as React from "react";
+import { Link, Routes, Route } from "react-router-dom";
 import { styled } from "@mui/material/styles";
-import { Link as Scroll } from "react-scroll";
 import Box from "@mui/material/Box";
 import Drawer from "@mui/material/Drawer";
 import CssBaseline from "@mui/material/CssBaseline";
 import List from "@mui/material/List";
-import Typography from "@mui/material/Typography";
 import Divider from "@mui/material/Divider";
 import IconButton from "@mui/material/IconButton";
 import ChevronLeftIcon from "@mui/icons-material/ChevronLeft";
@@ -16,19 +15,15 @@ import ListItemIcon from "@mui/material/ListItemIcon";
 import ListItemText from "@mui/material/ListItemText";
 import ArticleIcon from "@mui/icons-material/Article";
 import HomeIcon from "@mui/icons-material/Home";
-import Container from "@mui/material/Container";
 import {
-  Summarization,
-  QuestionAnswering,
-  Classification,
-  Insertion,
-  CodeGeneration,
-  Reasoning,
-  AbstractInfo,
-  RollPlayConsultant,
   Navbar,
 } from "../components/index";
-
+import {
+  HomeAbstract,
+  BasicPrompts,
+  CodePrompts,
+  ConversationPrompts,
+} from "./index";
 const drawerWidth = 280;
 
 const MainBox = styled("main", {
@@ -66,22 +61,20 @@ export default function PromptDashboard() {
   const toggleDrawer = () => {
     setOpen(!open);
   };
-  const menuList = [
-    "summarization",
-    "questionAnswering",
-    "classification",
-    "insertion",
-    "codeGeneration",
-    "reasoning",
-    "rollPlayConsultant",
+  const menuItems = [
+    {
+      name: "BasicPrompts",
+      url: "/gpt-prompt-templates/basic/"
+    },
+    {
+      name: "CodePrompts",
+      url: "/gpt-prompt-templates/code/"
+    },
+    {
+      name: "ConversationPrompts",
+      url: "/gpt-prompt-templates/conversation/"
+    },
   ];
-  const onClickItem = (textId) => {
-    //指定位置までスクロールする
-    document
-      .getElementById(textId)
-      .scrollIntoView({ behavior: "smooth", block: "end", inline: "nearest" });
-    setOpen(false);
-  };
   return (
     <Box sx={{ display: "flex" }}>
       <CssBaseline />
@@ -108,70 +101,40 @@ export default function PromptDashboard() {
         </DrawerHeader>
         <Divider />
         <List>
-          <Scroll to={"top"} smooth>
-            <ListItem key="Top" disablePadding>
-              <ListItemButton onClick={() => onClickItem("top")}>
+          <ListItem key="Top" disablePadding>
+            <Link to="/gpt-prompt-templates/" >
+              <ListItemButton>
                 <ListItemIcon>
                   <HomeIcon />
                 </ListItemIcon>
                 <ListItemText primary="Top" />
               </ListItemButton>
-            </ListItem>
-          </Scroll>
+            </Link>
+          </ListItem>
         </List>
         <Divider />
         <List>
-          {menuList.map((text, index) => (
-            <ListItem key={text} disablePadding>
-              <ListItemButton onClick={() => onClickItem(text)}>
-                <ListItemIcon>
-                  <ArticleIcon />
-                </ListItemIcon>
-                <ListItemText primary={`${(index + 1)}.${text}`} />
-              </ListItemButton>
+          {menuItems.map((menu, index) => (
+            <ListItem key={index} disablePadding>
+              <Link to={menu.url} >
+                <ListItemButton>
+                  <ListItemIcon>
+                    <ArticleIcon />
+                  </ListItemIcon>
+                  <ListItemText primary={`${(index + 1)}.${menu.name}`} />
+                </ListItemButton>
+              </Link>
             </ListItem>
           ))}
         </List>
       </Drawer>
       <MainBox open={open}>
-        <DrawerHeader />
-        <header title="gpt-prompt-templates" id="top">
-          プロンプトテンプレート
-        </header>
-        <Typography
-          variant="h5"
-          gutterbottom="true"
-        >
-          プロンプト生成を補助するアプリです：
-        </Typography>
-        <Typography paragraph>
-          左メニューのテンプレートを選び独自の入力をすることで、プロンプトに貼り付け可能な文章を表示します。
-          テンプレートの解説を読んで、ご利用ください。
-        </Typography>
-        <Container>
-          <AbstractInfo title="Abstract" />
-        </Container>
-        <Container id="summarization">
-          <Summarization title="Summarization" />
-        </Container>
-        <Container id="questionAnswering">
-          <QuestionAnswering title="QuestionAnswering" />
-        </Container>
-        <Container id="classification">
-          <Classification title="Classification" />
-        </Container>
-        <Container id="insertion">
-          <Insertion title="Insertion" />
-        </Container>
-        <Container id="codeGeneration">
-          <CodeGeneration title="CodeGeneration" />
-        </Container>
-        <Container id="reasoning">
-          <Reasoning title="Reasoning" />
-        </Container>
-        <Container id="rollPlayConsultant">
-          <RollPlayConsultant title="RollPlayConsultant" />
-        </Container>
+        <Routes>
+          <Route path="/gpt-prompt-templates/" element={<HomeAbstract />} />
+          <Route path="/gpt-prompt-templates/basic/" element={<BasicPrompts />} />
+          <Route path="/gpt-prompt-templates/code/" element={<CodePrompts />} />
+          <Route path="/gpt-prompt-templates/conversation/" element={<ConversationPrompts />} />
+        </Routes>
       </MainBox>
     </Box>
   );
