+++ src/pages/index.js
@@ -1,2 +1,7 @@
-export { default as PromptDashboard } from './PromptDashboard'
+export { default as PromptDashboard } from './PromptDashboard';
+//
+export { default as HomeAbstract } from './HomeAbstract';
+export { default as BasicPrompts } from './BasicPrompts';
+export { default as CodePrompts } from './CodePrompts';
+export { default as ConversationPrompts } from './ConversationPrompts'
 //
